Line chart cleanup
