+++ locust/webui/src/components/LineChart/LineChart.constants.ts
@@ -1,0 +1,14 @@
+export const CHART_THEME = {
+  DARK: {
+    textColor: '#b3c3bc',
+    axisColor: '#5b6f66',
+    backgroundColor: '#27272a',
+    splitLine: '#4c4c52',
+  },
+  LIGHT: {
+    textColor: '#000',
+    axisColor: '#000',
+    backgroundColor: '#fff',
+    splitLine: '#ddd',
+  },
+};
